Use Integer IDs instead of Long/UUID
Convert primary key types across the codebase from Long/UUID to Integer. Updated entities (Component, EvaluationPlan, Management, Semester, Subject) to use Integer IDs, adjusted DTOs (ManagementResponse, SemesterRequest, SemesterResponse), repositories (JpaRepository generics and query signatures), controllers (path variable types) and services (method signatures and internal lookups). This aligns ID types with integer/serial DB columns (e.g., Postgres SERIAL) and removes UUID generation for Subject and EvaluationPlan. Note: database schema/migrations and any external clients must be updated to match the new integer IDs to avoid runtime issues.

+++ src/main/java/com/universidad_nur/notasnurv3_api/controllers/ManagementController.java
@@ -36,20 +36,20 @@ public class ManagementController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse<ManagementResponse>> getById(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<ManagementResponse>> getById(@PathVariable Integer id) {
         return ResponseEntity.ok(new ApiResponse<>(true, "Gestión encontrada", managementService.getById(id)));
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<ApiResponse<ManagementResponse>> update(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @Valid @RequestBody ManagementRequest request) {
         return ResponseEntity.ok(new ApiResponse<>(true, "Gestión actualizada", managementService.update(id, request)));
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<ApiResponse<Void>> deleteManagement(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<Void>> deleteManagement(@PathVariable Integer id) {
         managementService.deleteManagement(id);
         return ResponseEntity.ok(new ApiResponse<>(true, "Gestión eliminada correctamente", null));
     }
