+++ src/main/java/com/universidad_nur/notasnurv3_api/dto/SemesterResponse.java
@@ -3,11 +3,11 @@ package com.universidad_nur.notasnurv3_api.dto;
 import java.time.LocalDate;
 
 public record SemesterResponse(
-        Long id,
+        Integer id,
         Integer number,
         LocalDate startDate,
         LocalDate endDate,
-        Long managementId,
+        Integer managementId,
         Integer managementYear
 ) {
 }
